Backend/requirements.txt
fastapi
uvicorn
pydantic
ultralytics
opencv-python
python-multipart
python-jose[cryptography]
passlib[bcrypt]
requests
sqlalchemy
databases
asyncpg
python-dotenv
psycopg2-binary
